Refactor mixins into separate classes
