+++ src/main/java/com/github/quaoz/betterleads/mixin/AmbientEntityMixin.java
@@ -1,0 +1,26 @@
+package com.github.quaoz.betterleads.mixin;
+
+
+import com.github.quaoz.betterleads.BetterLeads;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.mob.AmbientEntity;
+import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+
+// Allows ambient entities (bats) to be leashed
+@Mixin(AmbientEntity.class)
+abstract class AmbientEntityMixin extends MobEntity {
+	protected AmbientEntityMixin(EntityType<? extends AmbientEntity> entityType, World world) {
+		super(entityType, world);
+	}
+	
+	@Inject(method = "canBeLeashedBy", at = @At("RETURN"), cancellable = true)
+	private void onCanBeLeashedBy(CallbackInfoReturnable<Boolean> cir) {
+		cir.setReturnValue((cir.getReturnValue() || (!this.isLeashed()) && BetterLeads.get().config.getLeashableAmbientMobs()));
+	}
+}
