+++ src/main/java/com/github/quaoz/betterleads/mixin/WaterCreatureEntityMixin.java
@@ -1,0 +1,26 @@
+package com.github.quaoz.betterleads.mixin;
+
+
+import com.github.quaoz.betterleads.BetterLeads;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.mob.PathAwareEntity;
+import net.minecraft.entity.mob.WaterCreatureEntity;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+
+// Allows water creatures (fish) to be leashed
+@Mixin(WaterCreatureEntity.class)
+abstract class WaterCreatureEntityMixin extends PathAwareEntity {
+	protected WaterCreatureEntityMixin(EntityType<? extends WaterCreatureEntity> entityType, World world) {
+		super(entityType, world);
+	}
+	
+	@Inject(method = "canBeLeashedBy", at = @At("RETURN"), cancellable = true)
+	private void onCanBeLeashedBy(CallbackInfoReturnable<Boolean> cir) {
+		cir.setReturnValue((cir.getReturnValue() || (!this.isLeashed()) && BetterLeads.get().config.getLeashableWaterCreatures()));
+	}
+}
