+++ src/main/resources/betterleads.mixins.json
@@ -1,9 +1,9 @@
 {
-  "required": true,
-  "minVersion": "0.8",
-  "package": "com.github.quaoz.betterleads.mixin",
+  "required":           true,
+  "minVersion":         "0.8",
+  "package":            "com.github.quaoz.betterleads.mixin",
   "compatibilityLevel": "JAVA_17",
-  "mixins": [
+  "mixins":             [
     "AmbientEntityMixin",
     "EndermanTeleportMixin",
     "MerchantEntityMixin",
@@ -12,8 +12,8 @@
     "TurtleEntityMixin",
     "WaterCreatureEntityMixin"
   ],
-  "client": [],
-  "injectors": {
+  "client":             [ ],
+  "injectors":          {
     "defaultRequire": 1
   }
 }
